# backend/static/openapi.yaml
openapi: "3.0.3"
info:
  title: "IOT Smart Home for Elderly — Alfred API"
  description: |
    REST API cho hệ thống nhà thông minh hỗ trợ người cao tuổi.
    
    ## Xác thực
    Tất cả endpoint (trừ `/api/auth/login`, `/api/auth/register`) đều yêu cầu signed bearer token.
    Token được ký bằng `itsdangerous` và được gửi qua **HttpOnly Cookie** (`batman_os_auth`) hoặc header `Authorization: Bearer <token>`.
    
    ## Rate Limiting
    - Login: 5 req/phút, 20 req/giờ
    - AI Chat: 30 req/phút, 300 req/giờ
    - Devices: 60 req/phút, 2000 req/giờ
    - Mặc định: 100 req/giờ, 500 req/ngày
  version: "2.0.0"
  contact:
    name: "Tu Dinh Huy — ITITIU21213"
    email: "ITITIU21213@student.hcmiu.edu.vn"
  license:
    name: "Academic (Thesis Project)"
servers:
  - url: "http://127.0.0.1:5000"
    description: "Local development"
  - url: "https://your-domain.com"
    description: "Production"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: batman_os_auth
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Authentication required"
...
    Device:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Đèn phòng khách"
        code:
          type: string
          example: "den"
        icon:
          type: string
          example: "💡"
        category:
          type: string
          enum: [light, fan, sensor, ac, other]
          example: "light"
        room_id:
          type: integer
          nullable: true
          example: 2
        status:
          type: object
          properties:
            state:
              type: string
              enum: [on, off, unknown]
              example: "off"
            updated_at:
              type: string
              format: date-time
        map_x:
          type: number
          nullable: true
        map_y:
          type: number
          nullable: true
    Room:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Phòng ngủ"
        user_id:
          type: integer
          example: 1
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/devices:
    get:
      tags: [devices]
      summary: Lấy danh sách thiết bị
      parameters:
        - in: query
          name: room_id
          schema:
            type: integer
          description: Lọc theo phòng
      responses:
        "200":
          description: Danh sách thiết bị
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: string
                    example: "online"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [devices]
      summary: Tạo thiết bị mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name:
                  type: string
                  example: "Đèn phòng ngủ"
                code:
                  type: string
                  example: "den_pn"
                  description: Mã duy nhất của thiết bị, phải khớp với firmware ESP32
                icon:
                  type: string
                  example: "💡"
                category:
                  type: string
                  enum: [light, fan, sensor, ac, other]
                  example: "light"
                room_id:
                  type: integer
                  nullable: true
                control_types:
                  type: array
                  items:
                    type: string
                  example: ["on", "off"]
      responses:
        "201":
          description: Thiết bị đã được tạo
        "400":
          description: Dữ liệu không hợp lệ
...
  /api/rooms:
    get:
      tags: [rooms]
      summary: Lấy danh sách phòng
      responses:
        "200":
          description: Danh sách phòng
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Room"

    post:
      tags: [rooms]
      summary: Tạo phòng mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "Phòng ngủ"
      responses:
        "201":
          description: Phòng đã được tạo
...
components:
  responses:
    Unauthorized:
      description: Token không hợp lệ hoặc hết hạn
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"